Fix ScoreController & addScore Via Tool
