+++ src/EduManagementLab.Api/Controllers/LTIResultsController.cs
@@ -1,9 +1,11 @@
 using EduManagementLab.Core.Services;
-using EduManagementLab.IdentityServer4.Validation;
+using LtiAdvantage;
 using LtiAdvantage.AssignmentGradeServices;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Newtonsoft.Json;
+using System.ComponentModel.DataAnnotations;
 
 namespace EduManagementLab.Api.Controllers
 {
@@ -43,10 +45,89 @@ namespace EduManagementLab.Api.Controllers
                     ResultMaximum = 100,
                     ResultScore = (double)courseTaskResult.Score,
                     UserId = courseTaskResult.Membership.UserId.ToString(),
-                    ScoreOf = $"https://localhost:7134/LTILineItems/{courseId}/LTILineItem/{courseTaskResult.Id}"
+                    ScoreOf = $"https://localhost:7134/LTILineItems/{courseId}/LTILineItem/{courseTaskResult.Id}",
                 });
             }
             return lineItemsResults;
         }
+
+        [HttpPost]
+        [Consumes(Constants.MediaTypes.Score)]
+        [Produces(Constants.MediaTypes.Score)]
+        [ProducesResponseType(typeof(Score), StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = Constants.LtiScopes.Ags.Score)]
+        [Route("{courseId}/LTILineItem/{LTILineItemId}/scores", Name = Constants.ServiceEndpoints.Ags.ScoresService)]
+        public ActionResult<Score> AddLTIScore(Guid courseId, Guid LTILineItemId, object score)
+        {
+            //var courseTaskResult = _courseService.GetCourse(courseId, true).CourseTasks.FirstOrDefault(c => c.Id == LTILineItemId).Results.FirstOrDefault(c => c.MembershipId == Guid.Parse(score.UserId));
+
+
+            ////courseTaskResult.Score = (decimal)score.ScoreGiven;
+
+            Score scores = JsonConvert.DeserializeObject<Score>(score.ToString());
+
+
+            var memberId = _courseService.GetCourse(courseId, true).Memperships.FirstOrDefault(u => u.UserId == Guid.Parse(scores.UserId)).Id;
+
+            var updatedScore = _courseTaskService.UpdateCourseTaskResult(LTILineItemId, memberId, (decimal)scores.ScoreGiven);
+
+            var newScore = new Score
+            {
+                ScoreGiven = (double)updatedScore.Score,
+                Comment = updatedScore.Score > 75 ? "Good job!" : "Work harder..",
+                UserId = updatedScore.MembershipId.ToString(),
+                ScoreMaximum = 100,
+                ActivityProgress = ActivityProgress.Submitted,
+                GradingProgress = GradingProgess.FullyGraded,
+                TimeStamp = DateTime.Now,
+            };
+
+            //var lineItemsResults = new List<Score>();
+            //Score scoreResult = new Score();
+            //scoreResult = new Score
+            //{
+            //    UserId = courseTaskResult.MembershipId.ToString(),
+            //    ScoreGiven = score,
+            //    ScoreMaximum = 100,
+            //    Comment = courseTaskResult.Score > 60 ? "Good job!" : "Work harder..",
+            //    ActivityProgress = ActivityProgress.InProgress,
+            //    GradingProgress = GradingProgess.Pending,
+            //    TimeStamp = DateTime.Now
+            //};
+            //var url = $"https://localhost:7134/LTIResults/{courseId}/LTILineItem/{LTILineItemId}/results/{updatedScore.Id}";
+
+            //// Save the score
+            //return Created(url, newScore);
+
+            return newScore;
+        }
+
+        [HttpGet]
+        [Produces(Constants.MediaTypes.Score)]
+        [ProducesResponseType(typeof(Score), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = Constants.LtiScopes.Ags.ScoreReadonly + " " + Constants.LtiScopes.Ags.Score)]
+        [Route("{courseId}/LTILineItem/{LTILineItemId}/scores/{scoreId}")]
+        public ActionResult<Score> GetLTIScore(Guid courseId, Guid LTILineItemId, Guid memberId)
+        {
+            var courseTaskResult = _courseService.GetCourse(courseId, true).CourseTasks.FirstOrDefault(c => c.Id == LTILineItemId).Results.FirstOrDefault(c => c.MembershipId == memberId);
+
+            return new Score
+            {
+                UserId = courseTaskResult.MembershipId.ToString(),
+                ScoreGiven = (double)courseTaskResult.Score,
+                ScoreMaximum = 100,
+                Comment = courseTaskResult.Score > 60 ? "Good job!" : "Work harder..",
+                ActivityProgress = ActivityProgress.InProgress,
+                GradingProgress = GradingProgess.Pending,
+                TimeStamp = DateTime.Now
+            };
+        }
+
     }
 }
